// src/constants/enums.ts
export enum SeatType {
  WINDOW = 'window',
  MIDDLE = 'middle',
  AISLE = 'aisle',
}

export enum CabinClass {
  PRIVATE = 'private',
  FIRST = 'first',
  BUSINESS = 'business',
  ECONOMY_PLUS = 'economyPlus',
  ECONOMY = 'economy',
}

export enum FlightReason {
  OTHER = 'other',
  CREW = 'crew',
  BUSINESS = 'business',
  LEISURE = 'leisure',
}